chore: break up preview generation worker module
