+++ crates/server/src/lib.rs
@@ -4,6 +4,6 @@ pub mod env;
 pub mod utils;
 
 pub mod workers {
-    pub mod generate_previews;
+    pub mod previews;
 }
